restructured files, cleaned nodes package, fixed #3
- resturtured the files into dirs
- cleaned the whole nodes package (nodes.py)
- added demo_example2 project which is the original demo using the new nodes
- added gitignore
